+++ app/Http/Requests/Api/Auth/Login/StoreRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Api\Auth;
+namespace App\Http\Requests\Api\Auth\Login;
 
 use App\Http\Requests\ApiRequest;
 
-class LoginRequest extends ApiRequest
+class StoreRequest extends ApiRequest
 {
     /**
      * Get the validation rules that apply to the request.
