+++ app/Http/Requests/Api/Main/Concert/StoreRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Api\Transaction;
+namespace App\Http\Requests\Api\Main\Concert;
 
 use App\Http\Requests\ApiRequest;
 
-class UpdateRequest extends ApiRequest
+class StoreRequest extends ApiRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,10 +24,10 @@ class UpdateRequest extends ApiRequest
     public function rules()
     {
         return [
-            'concert' => ['nullable','numeric'],
-            'user' => ['nullable','numeric'],
-            'paid_at' => ['nullable','string','date_format:d-m-Y'],
-            'book_at' => ['nullable','string','date_format:d-m-Y']
+            'company' => ['required','numeric'],
+            'name' => ['required','string','max:255','unique:concerts,name'],
+            'start_at' => ['required','string','date_format:d-m-Y'],
+            'end_at' => ['required','string','date_format:d-m-Y']
         ];
     }
 }
