+++ app/Http/Requests/Api/Main/Concert/UpdateRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Api\Transaction;
+namespace App\Http\Requests\Api\Main\Concert;
 
 use App\Http\Requests\ApiRequest;
 
-class StoreRequest extends ApiRequest
+class UpdateRequest extends ApiRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,10 +24,10 @@ class StoreRequest extends ApiRequest
     public function rules()
     {
         return [
-            'concert' => ['required','numeric'],
-            'user' => ['required','numeric'],
-            'paid_at' => ['required','string','date_format:d-m-Y'],
-            'book_at' => ['required','string','date_format:d-m-Y']
+            'company' => ['nullable','numeric'],
+            'name' => ['nullable','string','max:255','unique:concerts,name'],
+            'start_at' => ['nullable','string','date_format:d-m-Y'],
+            'end_at' => ['nullable','string','date_format:d-m-Y']
         ];
     }
 }
